pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "panel-model-viewer"
version = "0.1.0"
description = "A Panel component for the <model-viewer> web component."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "panel>=1.4.0",
    "param>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-playwright>=0.4.0",
    "ruff>=0.3.0",
    "ty>=0.0.1", 
    "watchfiles>=0.21.0",
]

[tool.hatch.build.targets.wheel]
packages = ["panel_model_viewer"]
include = ["panel_model_viewer/static/*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
markers = [
    "ui: marks tests as UI tests (deselect with '-m \"not ui\"')",
]

[tool.ty]
# Configuration for ty (if needed, otherwise defaults)
